% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_utils.R
\name{add_dependencies_to_description}
\alias{add_dependencies_to_description}
\title{Add a "Depends:"-field to the DESCRIPTION file}
\usage{
add_dependencies_to_description(package_directory, dependencies = NULL)
}
\arguments{
\item{package_directory}{Path to the directory.}

\item{dependencies}{the package names the temporary package will depend on.}
}
\value{
value of \code{link{writeLines}}.
}
\description{
if the functions in the temporary package depend on other functions (from,
for example, the checkmate package), you can add the whole package as a
dependency.
}
\section{Version}{
 $Id: 3d72fa8d41ea4e2e7b29ffdfefc45e58fe5274b5 $
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}

