% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{alter_description}
\alias{alter_description}
\title{Alter Contents of a DESCRIPTION File}
\usage{
alter_description(path, substitution = NULL, addition = NULL)
}
\arguments{
\item{path}{Path to the DESCRIPTION file or the directory containing it.}

\item{substitution}{A list of named character vector giving the pairs for
substitution.}

\item{addition}{A list of named character vector giving the pairs for
addition}
}
\value{
value of \code{\link{write.dcf}}.
}
\description{
Substitutes and/or adds fields in a DESCRIPTION file.
}
