% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{alter_description_file}
\alias{alter_description_file}
\title{Alter Contents of a DESCRIPTION File}
\usage{
alter_description_file(path, substitution = NULL, addition = NULL)
}
\arguments{
\item{path}{Path to the DESCRIPTION file or the directory containing it.}

\item{substitution}{A list of named character vector giving the pairs for
substitution.}

\item{addition}{A list of named character vector giving the pairs for
addition.}
}
\value{
value of \code{\link{write.dcf}}.
}
\description{
Substitutes and/or adds fields in a DESCRIPTION file.
}
\note{
Adding NULL elements (Title = NULL, for example) to substitution
does not delete or overwrite the entry in the DESCRIPTION file. \cr
\emph{alter_description_file() is deprecated, please use CRAN package
\pkg{desc}}.
}
\keyword{internal}
