% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_utils.R
\name{build_and_check_package}
\alias{build_and_check_package}
\title{build and check a temporary package intended for documentation mainly.}
\usage{
build_and_check_package(package_directory,
  copy_tmp_files_to = dirname(tempdir()))
}
\arguments{
\item{package_directory}{Path to the packages' directory.}

\item{copy_tmp_files_to}{Path to save output from R CMD check to.}
}
\value{
invisibly 0, if R CMD check returned 0, nothing otherwise.
}
\description{
We might want to see if the temporary package we've created to document our
code file would build and pass R CMD check as a proper package.
}
\section{Version}{
 $Id: 3d72fa8d41ea4e2e7b29ffdfefc45e58fe5274b5 $
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}

