% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{check_package}
\alias{check_package}
\title{Run \command{R CMD check} on a Package Directory}
\usage{
check_package(path, working_directory, as_cran,
  throw_on_check_not_passing = TRUE)
}
\arguments{
\item{path}{The path to the package directory.}

\item{working_directory}{A working directory. Keep the default.}

\item{as_cran}{Use the \command{--as-cran} flag?}

\item{throw_on_check_not_passing}{Stop if \command{R CMD check} does not
pass?}
}
\value{
The return value of \command{R CMD check} ran through
\code{callr::rcmd_safe()}.
}
\description{
Run \command{R CMD check} on a Package Directory
}
