% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/man.R
\name{display_Rd}
\alias{display_Rd}
\title{Display the Contents of an R documentation file}
\usage{
display_Rd(rd_file)
}
\arguments{
\item{rd_file}{The path to the Rd file to be displayed.}
}
\value{
the return value of removing the temporary file.
}
\description{
This is meant for internal use by \code{\link{man}}.
}
\note{
The Contents are converted to text with \code{Rdconv} and then saved
to a temporary file which is then displayed using the R pager.
Using \code{\link{cat}} on the text would not allow for using different
pagers.
}
\examples{
document(file_name = system.file("tests", "files", "simple.R", package = "document"),
         check_package = FALSE)
document:::display_Rd(list.files(file.path(get_dpd(), "man"), full.names = TRUE)[2])
}
\author{
Andreas Dominik Cullmann, <adc-r@arcor.de>
}
