% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{document}
\alias{document}
\title{roxygenize an R code file, output the documentation to pdf.}
\usage{
document(file_name, working_directory = NULL, output_directory = tempdir(),
  dependencies = NULL, sanitize_Rd = TRUE, runit = FALSE,
  check_package = TRUE, clean = FALSE, ...)
}
\arguments{
\item{file_name}{The name of the R code file to be documented.}

\item{working_directory}{A working directory. Keep the default.}

\item{output_directory}{The directory to put the documentation into.}

\item{dependencies}{a character vector of package names the functions depend
on.}

\item{sanitize_Rd}{Remove strange characters from \code{Rdconv}?
Do not bother, this is for Unit testing only.
on.}

\item{runit}{Convert the text received from the help files if running RUnit?
Do not bother, this is for Unit testing only.
on.}

\item{check_package}{Run \code{\link[devtools]{check}} on the sources? Do not
forget to export your functions if set to TRUE.}

\item{clean}{Delete the working directory?}

\item{...}{Arguments passed to \code{\link{get_lines_between_tags}}.}
}
\value{
A list containing
\describe{
    \item{pdf_path}{The path to the pdf file produced.}
    \item{txt_path}{The path to the text file produced.}
    \item{html_path}{The path to the html file produced.}
    \item{check_result}{A list giving the R CMD check results.}
}
}
\description{
roxygenize an R code file, output the documentation to pdf.
}
\examples{
res <- document(file_name = system.file("tests", "files", "minimal.R",
                                        package = "document"),
                check_package = TRUE)
cat(res[["check_result"]][["stdout"]], sep = "\\n")
cat(res[["check_result"]][["stderr"]], sep = "\\n")
}
\author{
Andreas Dominik Cullmann, <adc-r@arcor.de>
}
