% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{get_dpd}
\alias{get_dpd}
\title{A Convenience Wrapper}
\usage{
get_dpd()
}
\value{
\code{getOption("document_package_directory")}
}
\description{
Just a wrapper to \code{\link[base]{getOption}("document_package_directory")}
}
