% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{get_missing_packages}
\alias{get_missing_packages}
\title{check which of the required packages are not installed on the system}
\usage{
get_missing_packages(required_packages)
}
\arguments{
\item{required_packages}{a vector of names of the required packages.}
}
\value{
a vector of names of the missing packages.
}
\description{
check which of the required packages are not installed on the system
}
\section{Version}{
 $Id: ab696f28eb22651bc1bfc8772d53d777927e82db $
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}

