% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_pandoc}
\alias{install_pandoc}
\title{install pandoc}
\usage{
install_pandoc()
}
\value{
TODO
}
\description{
pandoc is provided via installr::install.pandoc
}
\section{Version}{
 $Id: ab696f28eb22651bc1bfc8772d53d777927e82db $
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}

