% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{is_pandoc_installed}
\alias{is_pandoc_installed}
\title{test if pandoc is installed on the system}
\usage{
is_pandoc_installed()
}
\value{
TRUE if pandoc is installed, FALSE otherwise
}
\description{
test if pandoc is installed on the system
}
\section{Version}{
 $Id: ab696f28eb22651bc1bfc8772d53d777927e82db $
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}

