% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/man.R
\name{is_Rd_file}
\alias{is_Rd_file}
\title{XXX:}
\usage{
is_Rd_file(x)
}
\arguments{
\item{x}{XXX:}
}
\value{
XXX:
}
\description{
XXX:
}
\examples{
warning("non given")
}
\author{
Andreas Dominik Cullmann, <adc-r@arcor.de>
}
