% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/man.R
\name{man}
\alias{man}
\title{XXX: view a help page from a file documentation}
\usage{
man(x, topic = NA, force_Rd = FALSE)
}
\arguments{
\item{x}{A XXX:}

\item{topic}{XXX:}

\item{force_Rd}{XXX:}
}
\value{
Invisibly XXX:
}
\description{
XXX: view a help page from a file documentation
}
\examples{
document(file_name = system.file("tests", "files", "simple.R", package = "document"), 
         check_package = FALSE)
man("a_first_function")
}
\author{
Andreas Dominik Cullmann, <adc-r@arcor.de>
}
