% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{provide_packages}
\alias{provide_packages}
\title{provide required packages}
\usage{
provide_packages(packages, mirror = "http://ftp5.gwdg.de/pub/misc/cran/")
}
\arguments{
\item{packages}{a vector of names of the required packages.}

\item{mirror}{the repository mirror to be used for installation of missing
packages.}
}
\value{
TRUE if all required packages were succesfully loaded, FALSE
otherwise.
}
\description{
install packages as required, then load all required packages.
}
\section{Version}{
 $Id: ab696f28eb22651bc1bfc8772d53d777927e82db $
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}

