% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_utils.R
\name{remove_package_Rd}
\alias{remove_package_Rd}
\title{Remove autogenerated package Rd from man directory}
\usage{
remove_package_Rd(package_directory)
}
\arguments{
\item{package_directory}{Path to the directory.}
}
\value{
value of \code{link{base::file.remove()}}.
}
\description{
utils::package.skeleton() and roxygen2::roxygenize() leave us with an invalid
PACKAGENAME-package.Rd file in the man directory. It needs to be removed.
}
\section{Version}{
 $Id: 3d72fa8d41ea4e2e7b29ffdfefc45e58fe5274b5 $
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}

