% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helpers.R
\name{set_RUN_R_TESTS}
\alias{set_RUN_R_TESTS}
\title{Set System Variable RUN_R_TESTS}
\usage{
set_RUN_R_TESTS(x, force = FALSE)
}
\arguments{
\item{x}{A logical, typically some function output.}

\item{force}{Overwrite the variable if already set?}
}
\value{
The value RUN_R_TESTS is set to, \code{\link{NULL}} nothing is done.
}
\description{
Set System Variable RUN_R_TESTS
}
\examples{
set_RUN_R_TESTS(is_running_on_fvafrcu_machines())
get_RUN_R_TESTS()
set_RUN_R_TESTS(TRUE, force = TRUE)
get_RUN_R_TESTS()
}
\keyword{internal}
