% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/man.R
\name{show_Rd}
\alias{show_Rd}
\title{XXX:}
\usage{
show_Rd(rd_file)
}
\arguments{
\item{rd_file}{XXX:}
}
\value{
XXX:
}
\description{
XXX:
}
\examples{
warning("non given")
}
\author{
Andreas Dominik Cullmann, <adc-r@arcor.de>
}
