% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_utils.R
\name{sort_unlocale}
\alias{sort_unlocale}
\title{sort a character vector in lexical order avoiding the locale.}
\usage{
sort_unlocale(char)
}
\arguments{
\item{char}{The character vector to sort.}
}
\value{
The sorted character vector.
}
\description{
\code{\link{sort}} uses the collation of the locale (see
\code{\link{Comparison}}), which results in different sortings. If a (RUnit)
check relies on sorting, we need to avoid the locale.
}
\section{Version}{
 $Id: 3d72fa8d41ea4e2e7b29ffdfefc45e58fe5274b5 $
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}

