% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/man.R
\name{usage}
\alias{usage}
\title{Return the Usage of a Function From Within the Function}
\usage{
usage()
}
\value{
A character string giving the Usage as \code{\link{help}} would do.
}
\description{
Return the Usage of a Function From Within the Function
}
\note{
Do not encapsulate it inside a function like \code{\link{print}} or
\code{\link{assign}} since it will then return that function's Usage. Use the
\code{<-} assignment operator.
}
\examples{
foo <- function(x){
    u <- usage()
    message("Usage is: ", u)
}
foo()
}
\author{
Andreas Dominik Cullmann, <adc-r@arcor.de>
}
