% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{write_the_docs}
\alias{write_the_docs}
\title{Read R Documentation Files from a Package's Source, Convert and Write Them
to Disk}
\usage{
write_the_docs(package_directory, file_name, output_directory = tempdir(),
  dependencies = NULL, sanitize_Rd = TRUE, runit = FALSE)
}
\arguments{
\item{package_directory}{The directory containing the package's source.}

\item{file_name}{The name of the R code file to be documented.}

\item{output_directory}{The directory to put the documentation into.}

\item{dependencies}{a character vector of package names the functions depend
on.}

\item{sanitize_Rd}{Remove strange characters from \code{Rdconv}?
Do not bother, this is for Unit testing only.
on.}

\item{runit}{Convert the text received from the help files if running RUnit?}
}
\value{
A list containing
\describe{
    \item{pdf_path}{The path to the pdf file produced.}
    \item{txt_path}{The path to the text file produced.}
    \item{html_path}{The path to the html file produced.}
}
}
\description{
Read R Documentation Files from a Package's Source, Convert and Write Them
to Disk
}
\author{
Andreas Dominik Cullmann, <adc-r@arcor.de>
}
