% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_dependencies_to_description}
\alias{add_dependencies_to_description}
\title{Add a "Depends:"-field to the DESCRIPTION File}
\usage{
add_dependencies_to_description(package_directory, dependencies = NULL)
}
\arguments{
\item{package_directory}{Path to the directory.}

\item{dependencies}{the package names the temporary package will depend on.}
}
\value{
value of \code{link{writeLines}}.
}
\description{
if the functions in the temporary package depend on other functions (from,
for example, the checkmate package), you can add the whole package as a
dependency.
}
\author{
Andreas Dominik Cullmann, <adc-r@arcor.de>
}
