% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_dependencies_to_description}
\alias{add_dependencies_to_description}
\title{Add a Dependency to the DESCRIPTION File}
\usage{
add_dependencies_to_description(package_directory, dependencies = NULL)
}
\arguments{
\item{package_directory}{The directory containing the package's source.}

\item{dependencies}{the package names the temporary package will depend on.}
}
\value{
value of \code{\link{writeLines}}.
}
\description{
If the functions in a package depend on other functions (from,
for example, the checkmate package), you can add the whole package as a
dependency.
}
