% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{check_package}
\alias{check_package}
\title{Run \command{R CMD check} on a Package Directory}
\usage{
check_package(package_directory, working_directory, check_as_cran = TRUE,
  stop_on_check_not_passing = TRUE, debug = TRUE)
}
\arguments{
\item{package_directory}{The directory containing the package's source.}

\item{working_directory}{A working directory. Keep the default.}

\item{check_as_cran}{Use the \command{--as-cran} flag with
\command{R CMD check}?}

\item{stop_on_check_not_passing}{Stop if \command{R CMD check} does not pass?}

\item{debug}{For internal use only: Summarize errors for \command{travis}?}
}
\value{
The return value of \command{R CMD check} ran through
\code{callr::rcmd_safe()}.
}
\description{
Run \command{R CMD check} on a Package Directory
}
