% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_description}
\alias{clean_description}
\title{Change License in the DESCRIPTION File to "GPL"}
\usage{
clean_description(package_directory)
}
\arguments{
\item{package_directory}{Path to the directory.}
}
\value{
value of \code{\link{writeLines}}.
}
\description{
\code{utils::\link[utils]{package.skeleton}} leaves us with a DESCRIPTION 
that throws a warning in \command{R CMD check}. Fix that.
}
\note{
We also set Version to 1.0.0.
}
