% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_description}
\alias{clean_description}
\title{Change License in the DESCRIPTION File to "GPL"}
\usage{
clean_description(package_directory)
}
\arguments{
\item{package_directory}{Path to the directory.}
}
\value{
value of \code{link{writeLines}}.
}
\description{
utils::package.skeleton() leaves us with a DESCRIPTION that throws a warning
in R CMD check. Fix that.
}
\note{
We also set Version to 0.0-0, hoping nobody is going to use
that.
}
\author{
Andreas Dominik Cullmann, <adc-r@arcor.de>
}
