% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/man.R
\name{display_Rd}
\alias{display_Rd}
\title{Display the Contents of an R Documentation File}
\usage{
display_Rd(rd_file)
}
\arguments{
\item{rd_file}{The path to the Rd file to be displayed.}
}
\value{
The return value of removing the temporary file.
}
\description{
This is meant for internal use by \code{\link{man}}.
}
\note{
The Contents are converted to text with \code{Rdconv} and then saved
to a temporary file which is then displayed using the R pager.
Using \code{\link{cat}} on the text would not allow for using different
pagers.
}
\keyword{internal}
