% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{document}
\alias{document}
\title{roxygenize an R code file, output the documentation to pdf.}
\usage{
document(file_name, output_directory = tempdir(), check_package = TRUE,
  working_directory = tempdir(), dependencies = NULL, runit = FALSE, ...)
}
\arguments{
\item{file_name}{The name of the R code file to be documented.}

\item{output_directory}{The directory to put the documentation into.}

\item{check_package}{Run \code{\link[devtools]{check}} on the sources?}

\item{working_directory}{A working directory. 
\bold{Warning} the working_directory will be recursively
\code{\link{unlink}}ed. You can erase your disk if you change the default!}

\item{dependencies}{a character vector of package names the functions depend
on.}

\item{runit}{Convert the text received from the help files if running RUnit?
Do not bother, this is for Unit testing only.}

\item{...}{Arguments passed to \code{\link{get_lines_between_tags}}.}
}
\value{
A list containing
\describe{
    \item{pdf_path}{The path to the pdf file produced.}
    \item{txt_path}{The path to the text file produced.}
    \item{check_result}{The value of \code{\link[devtools]{check}}.}
}
}
\description{
roxygenize an R code file, output the documentation to pdf.
}
\examples{
document(file_name = system.file("tests", "files", "simple.R", package = "document"),
         check_package = FALSE)
}
\author{
Andreas Dominik Cullmann, <adc-r@arcor.de>
}
