% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{document}
\alias{document}
\title{roxygenize an R code file, output the documentation to pdf.}
\usage{
document(file_name, output_directory = ".", clean = FALSE,
  check_package = TRUE, working_directory = tempfile(),
  dependencies = NULL, ...)
}
\arguments{
\item{file_name}{The name of the R code file to be documented.}

\item{output_directory}{The directory to put the documentation into.}

\item{clean}{Delete the working direcotry?}

\item{check_package}{Run R CMD check on the sources?}

\item{working_directory}{A working directory. Defaults to tempdir().
\bold{Warning} the working_directory will be recursively
\code{\link{unlink}}ed. You can erase your disk if you change the default!}

\item{dependencies}{a character vector of package names the functions depend
on.}

\item{...}{Arguments passed to \code{\link{get_lines_between_tags}}.}
}
\value{
TRUE if pdf creation is successfull, FALSE otherwise.
}
\description{
roxygenize an R code file, output the documentation to pdf.
}
\section{Version}{
 $Id: 9bbb752b06d887f2115e37c3e9dadd89e40c49c7 $
}
\examples{
document(file_name = system.file("tests", "files", "simple.R", package = "document"))
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}

