% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{document}
\alias{document}
\title{roxygenize an R code file, output the documentation to pdf.}
\usage{
document(file_name, output_directory = getwd(), clean = TRUE,
  check_package = TRUE, working_directory = tempfile(),
  dependencies = NULL, runit = FALSE, ...)
}
\arguments{
\item{file_name}{The name of the R code file to be documented.}

\item{output_directory}{The directory to put the documentation into.}

\item{clean}{Delete the working directory?}

\item{check_package}{Run R CMD check on the sources?}

\item{working_directory}{A working directory. 
\bold{Warning} the working_directory will be recursively
\code{\link{unlink}}ed. You can erase your disk if you change the default!}

\item{dependencies}{a character vector of package names the functions depend
on.}

\item{runit}{Convert the text recieved from the help files if running RUnit?
Do not bother, this is for Unit testing only.}

\item{...}{Arguments passed to \code{\link{get_lines_between_tags}}.}
}
\value{
FIXME
}
\description{
roxygenize an R code file, output the documentation to pdf.
}
\examples{
document(file_name = system.file("tests", "files", "simple.R", package = "document"))
}
\author{
Andreas Dominik Cullmann, <adc-r@arcor.de>
}
