% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{fake_package}
\alias{fake_package}
\title{Fake a Package From a Single Code File}
\usage{
fake_package(file_name, working_directory = NULL, dependencies = NULL, ...)
}
\arguments{
\item{file_name}{The name of the R code file to be documented.}

\item{working_directory}{A working directory. Keep the default.}

\item{dependencies}{A character vector of package names the functions depend
on.}

\item{...}{Arguments passed to \code{\link{get_lines_between_tags}}.}
}
\value{
A string containing the path to the faked package.
}
\description{
To build documentation for a single code file, we need a temporary package.
Of course the code file should contain \pkg{roxygen2} comments.
}
\examples{
fake_package(file_name = system.file("tests", "files", "simple.R",
         package = "document"))
}
