% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_utils.R
\name{fix_package_documentation}
\alias{fix_package_documentation}
\title{fix a \code{\link{package.skeleton}}s package skeleton.}
\usage{
fix_package_documentation(package_directory)
}
\arguments{
\item{package_directory}{Path to the directory.}
}
\value{
invisibly NULL.
}
\description{
This is just a convenience wrapper to \code{\link{remove_package_Rd}} and
\code{\link{clean_description}}.
}
\author{
Andreas Dominik Cullmann, <adc-r@arcor.de>
}

