% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{get_dpd}
\alias{get_dpd}
\title{A Convenience Wrapper to \code{getOption("document_package_directory")}}
\usage{
get_dpd()
}
\value{
\code{getOption("document_package_directory")}
}
\description{
A Convenience Wrapper to \code{getOption("document_package_directory")}
}
\author{
Andreas Dominik Cullmann, <adc-r@arcor.de>
}
