% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/man.R
\name{is_Rd_file}
\alias{is_Rd_file}
\title{Check Whether a File is Probably an R documentation file}
\usage{
is_Rd_file(x)
}
\arguments{
\item{x}{The path to the file to be checked.}
}
\value{
TRUE if the file is probably an R documentation file, FALSE
otherwise.
}
\description{
This is meant for internal use by \code{\link{man}}.
}
\note{
The check might produce false negatives (erroneously assuming the file
is not an R documentation file).
}
\author{
Andreas Dominik Cullmann, <adc-r@arcor.de>
}
