% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/man.R
\name{man}
\alias{man}
\title{view a help page from a file documentation}
\usage{
man(fun)
}
\arguments{
\item{fun}{A string giving the name of the function.}
}
\value{
Invisibly NULL
}
\description{
view a help page from a file documentation
}
\examples{
document(file_name = system.file("tests", "files", "simple.R", package = "document"), 
         check_package = FALSE)
man("a_first_function")
}
\author{
Andreas Dominik Cullmann, <adc-r@arcor.de>
}
