% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Rd_txt_RUnit}
\alias{Rd_txt_RUnit}
\title{clean a string created from a run through RUnit}
\usage{
Rd_txt_RUnit(txt)
}
\arguments{
\item{txt}{a character vector}
}
\value{
the sanitized character vector
}
\description{
Why am I doing this? It want to run RUnit tests from within R CMD check
and interactively. Files produced are compared with expected files. Now R
CMD check and interactive (and batch) give different encodings. I don't
know why, but they do.
}
\author{
Andreas Dominik Cullmann, <adc-r@arcor.de>
}
