% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_package_Rd}
\alias{remove_package_Rd}
\title{Remove Automatically Generated Package Rd from Man Directory}
\usage{
remove_package_Rd(package_directory)
}
\arguments{
\item{package_directory}{Path to the directory.}
}
\value{
value of \code{\link{file.remove}}.
}
\description{
\code{\link[utils]{package.skeleton}} and
\code{\link[roxygen2]{roxygenize}}
leave us with an invalid name-package.Rd file in the man directory. 
So we remove it.
}
\author{
Andreas Dominik Cullmann, <adc-r@arcor.de>
}
