% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_utils.R
\name{remove_package_Rd}
\alias{remove_package_Rd}
\title{Remove automatically generated package Rd from man directory}
\usage{
remove_package_Rd(package_directory)
}
\arguments{
\item{package_directory}{Path to the directory.}
}
\value{
value of \code{link{base::file.remove()}}.
}
\author{
Andreas Dominik Cullmann, <adc-r@arcor.de>
}

