% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_utils.R
\name{sort_unlocale}
\alias{sort_unlocale}
\title{sort a character vector in lexical order avoiding the locale.}
\usage{
sort_unlocale(char)
}
\arguments{
\item{char}{The character vector to sort.}
}
\value{
The sorted character vector.
}
\description{
\code{\link{sort}} uses the collation of the locale (see
\code{\link{Comparison}}), which results in different sorting. If a (RUnit)
check relies on sorting, we need to avoid the locale.
}
\author{
Andreas Dominik Cullmann, <adc-r@arcor.de>
}
